Modify code to remove all warnings

+++ include/EDF/src/Queue.tpp
@@ -97,7 +97,7 @@ template<typename T, std::size_t N>
 constexpr std::uint16_t Queue<T, N>::
 pop16be() {
     static_assert( std::is_same_v<T, std::uint8_t>, "pop16be() is only available if T is a 'std::uint8_t'");
-    uint16_t result = pop8be() << 8;
+    uint16_t result = static_cast<uint16_t>(static_cast<uint16_t>(pop8be()) << 8);
     result |= pop8be();
     return result;
 }
@@ -106,7 +106,7 @@ template<typename T, std::size_t N>
 constexpr std::uint32_t Queue<T,N>::
 pop32be() {
     static_assert( std::is_same_v<T, std::uint8_t>, "pop32be() is only available if T is a 'std::uint8_t'");
-    uint32_t result = pop16be() << 16;
+    uint32_t result = static_cast<uint32_t>(pop16be()) << 16;
     result |= pop16be();
     return result;
 }
@@ -141,7 +141,7 @@ constexpr std::uint32_t Queue<T,N>::
 pop32le() {
     static_assert( std::is_same_v<T, std::uint8_t>, "pop32le() is only available if T is a 'std::uint8_t'");
     uint32_t result = pop16le();
-    result |= pop16le() << 16;
+    result |= static_cast<uint32_t>(pop16le()) << 16;
     return result;
 }
